Implement While language and improve code
There is only one target evaluation language now and that is the
strict While language. Loop is simply translated to While and the
translated program is evaluated.

Also, the code has been improved considerably. I tried to extract
common functionality in order to reduce code duplication but I
could not abstract the code as much as I wanted even though I
literally thought about how to approach this problem half of the
time when working on this project.