Fix evaluation space leak
I encountered a problem when playing around with the evaluation service,
namely a stack overflow for this program:

    counter := 1;
    LOOP 1000 DO
      x0      := x0 + counter;
      counter := counter + 1
    END

After much investigation and profiling it turned out that the space leak
originated from unevaluated thunks of variable values. In fact, even if the
state in the ST.Monad is strict the values aren't necessarily. Putting two
$!'s in just the right places was the solution to the problem.