Remove Goto bug and remove Goto redundancy
I discovered an error in the reasoning for the transformation of extended Goto
to strict Goto which lead to simple statements like "x0 := x1 * x2" not working
in the Goto language. I promptly fixed the error.

At the same time I decided to reduce code redundancy in the Goto langauge
as the clutter of statements like "x0 := x0 + 0" annoyed me, although using
these statements for the code that is responsible for transforming extended
Goto to strict Goto made the code much easier. The output code is now free of
such statements.