Improve Goto evaluation speed drastically
Finally came around to write a dedicated evaluator for the Goto language as
the previous way of transforming a Goto Program into an equivalent While
program behind the curtains and then evaluating that While program turned out
to be too slow to be practical.

Basic Idea: Since a strict Goto Program is flat an array is created whose cells
are the respective Goto statements. If there is a jump in a Goto program the
evaluator simply accesses the statement at the specified index of the jump
(O(1)) and proceeds executing statements from there.

The way the evaluator is implemented now naturally maps to the Goto semantics
and thus, finally, the execution speed of Goto is practical. The icing on the
cake is that I can at last use the extended tests for the Goto language ;).