Improve evaluators (x9999999999 is no problem anymore)
The evaluators used to work with an infinite list of variable references as
the environment. However, this caused a problem when using a strict variable
identifier with a really big index like 'x999999999999'. The evaluator(s) had
to first create all intermediate variable references. That is a total of
1000000000000 (including 'x0') variable references which blew the stack.
Granted, using such a variable identifier is an artifical use case but if
someone wanted to kill the server of the live evaluation service easily he
could have simple entered 'x9999999999 := 0' as a program...

Now the evaluators simply use maps (Data.Map). Whenever a never before used
variable is requested a variable reference is created and put into the map.
The number of created variable references is now limited by the program's
length instead of the peculiarity of variable identifiers.

They lookup time of variables can probably be improved by using a hash table.