Improve Goto to strict Goto transformation
I observed that the toStrictLabels function in the to strict Goto
transformation code for the Goto language yielded a horrible performance for
source files with over a thousand lines (>3m -O2). By changing the algorithm
slightly, mainly changing the list difference computation, the performance has
been drastically improved (~14s).

However, the performance still was not as good as I wanted it to be. I found
a space leak whose origin I'm still not sure about (it only really appeared
if both removeRedundancy & toStrictLabels where present in succession in the
toStrict function). After a long battle I finally somehow closed the leak by
makind strict & extended Goto instances of NFData and using deepseq.

The time performance didn't improve however. It turns out that removeRedundancy
is causing a severe slowdown. I suspect it is the way I use renameGotoLabel
in foldr. I'm creating a lot of intermediate structures that are garbage
collected immediately (well, thanks to deepseq). Instead, I should write a
renameGotoLabels function that receives a list of remappings and only traverses
the abstract syntax tree once.

For now, I have created a toStrict' function that does not remove redundancy
and therefore is a lot faster.